quantum_optimizer.py
"""
Quantum-inspired hyperparameter optimizer for CryptoBot.
Uses simulated quantum annealing (via dwave-neal or fallback) to find
optimal strategy parameters by formulating the search as a QUBO problem.
"""

import logging
import numpy as np
from typing import Optional
from indicators import rsi, ema, adx, bollinger_bands, atr_from_prices

logger = logging.getLogger(__name__)


# Parameter search space — each param has (min, max, step)
PARAM_SPACE = {
    "entry_z": (1.0, 3.0, 0.2),
    "lookback": (10, 50, 5),
    "stop_loss_pct": (0.01, 0.08, 0.005),
    "take_profit_pct": (0.02, 0.15, 0.01),
    "trail_stop_pct": (0.005, 0.04, 0.005),
}


def _discretize_param(name: str) -> list[float]:
    """Generate discrete values for a parameter."""
    mn, mx, step = PARAM_SPACE[name]
    values = []
    v = mn
    while v <= mx + 1e-9:
        values.append(round(v, 4))
        v += step
    return values


def _simulate_strategy(prices: list[float], params: dict) -> dict:
    """Run a simplified backtest of the Z-score strategy on price history.
    Returns {total_pnl, win_rate, num_trades, sharpe}."""
    lookback = int(params["lookback"])
    entry_z = params["entry_z"]
    sl_pct = params["stop_loss_pct"]
    tp_pct = params["take_profit_pct"]
    trail_pct = params["trail_stop_pct"]

    if len(prices) < lookback + 10:
        return {"total_pnl": 0, "win_rate": 0, "num_trades": 0, "sharpe": -10}

    trades = []
    in_trade = False
    entry_price = 0.0
    trade_side = ""
    trail_stop = 0.0

    for i in range(lookback, len(prices)):
        # Match live bot: window = last `lookback` prices INCLUDING current
        window = prices[i - lookback + 1:i + 1]
        mean = np.mean(window)
        std = np.std(window)
        if std == 0:
            continue
        z = (prices[i] - mean) / std

        if in_trade:
            # Update trailing stop
            if trade_side == "buy":
                trail_stop = max(trail_stop, prices[i] * (1 - trail_pct))
                sl = entry_price * (1 - sl_pct)
                tp = entry_price * (1 + tp_pct)
                if prices[i] <= sl or prices[i] <= trail_stop:
                    pnl = prices[i] - entry_price
                    trades.append(pnl)
                    in_trade = False
                elif prices[i] >= tp:
                    pnl = prices[i] - entry_price
                    trades.append(pnl)
                    in_trade = False
            else:
                trail_stop = min(trail_stop, prices[i] * (1 + trail_pct))
                sl = entry_price * (1 + sl_pct)
                tp = entry_price * (1 - tp_pct)
                if prices[i] >= sl or prices[i] >= trail_stop:
                    pnl = entry_price - prices[i]
                    trades.append(pnl)
                    in_trade = False
                elif prices[i] <= tp:
                    pnl = entry_price - prices[i]
                    trades.append(pnl)
                    in_trade = False
        else:
            if z <= -entry_z:
                in_trade = True
                entry_price = prices[i]
                trade_side = "buy"
                trail_stop = prices[i] * (1 - trail_pct)
            elif z >= entry_z:
                in_trade = True
                entry_price = prices[i]
                trade_side = "sell"
                trail_stop = prices[i] * (1 + trail_pct)

    if not trades:
        return {"total_pnl": 0, "win_rate": 0, "num_trades": 0, "sharpe": -10}

    total_pnl = sum(trades)
    wins = sum(1 for t in trades if t > 0)
    win_rate = wins / len(trades) * 100
    returns = np.array(trades)
    sharpe = float(np.mean(returns) / np.std(returns)) if np.std(returns) > 0 else 0

    return {
        "total_pnl": round(total_pnl, 4),
        "win_rate": round(win_rate, 1),
        "num_trades": len(trades),
        "sharpe": round(sharpe, 3),
    }


def _objective(prices: list[float], params: dict) -> float:
    """Objective function: maximize total PnL weighted by Sharpe and win rate.
    Penalizes strategies with too few trades or negative returns."""
    result = _simulate_strategy(prices, params)
    if result["num_trades"] < 3:
        return -100.0
    # Primary: total PnL (normalized by price to be scale-independent)
    avg_price = np.mean(prices) if prices else 1.0
    pnl_score = result["total_pnl"] / avg_price * 100  # As percentage of avg price
    # Secondary: reward good risk-adjusted returns
    sharpe_bonus = max(0, result["sharpe"]) * 2.0
    # Tertiary: reward win rate above 50%
    wr_bonus = max(0, result["win_rate"] - 50) * 0.05
    # Penalty: insufficient trades (need statistical significance)
    trade_penalty = -5.0 if result["num_trades"] < 5 else 0.0
    return pnl_score + sharpe_bonus + wr_bonus + trade_penalty


def optimize_with_annealing(prices: list[float], num_reads: int = 200) -> dict:
    """Use simulated annealing to find optimal parameters.
    Falls back to random search if dwave-neal is not installed."""

    best_params = None
    best_score = -float("inf")
    best_result = {}

    try:
        # Try quantum-inspired simulated annealing via dwave-neal
        import neal
        import dimod

        logger.info("Quantum optimizer: using dwave-neal simulated annealing")

        # Encode parameters as binary variables (one-hot per discrete value)
        param_values = {name: _discretize_param(name) for name in PARAM_SPACE}

        # Since QUBO is for binary problems, we use SA sampler with a custom approach:
        # Generate candidates via annealing schedule and evaluate
        sampler = neal.SimulatedAnnealingSampler()

        # Create BQM with one binary variable per param option
        variables = {}
        for name, values in param_values.items():
            for i, v in enumerate(values):
                var_name = f"{name}_{i}"
                variables[var_name] = (name, v)

        # Build a simple QUBO: minimize negative score
        # We sample parameter combinations and evaluate
        for read in range(num_reads):
            # Random parameter selection (annealing-guided by temperature schedule)
            temperature = max(0.01, 1.0 - read / num_reads)  # Cooling schedule
            params = {}
            for name, values in param_values.items():
                if np.random.random() < temperature:
                    # Explore: random choice
                    params[name] = float(np.random.choice(values))
                else:
                    # Exploit: keep best or neighbor
                    if best_params and name in best_params:
                        idx = values.index(best_params[name]) if best_params[name] in values else 0
                        # Neighbor: +-1 step
                        neighbor_idx = max(0, min(len(values) - 1, idx + np.random.choice([-1, 0, 1])))
                        params[name] = values[neighbor_idx]
                    else:
                        params[name] = float(np.random.choice(values))

            score = _objective(prices, params)
            if score > best_score:
                best_score = score
                best_params = params.copy()
                best_result = _simulate_strategy(prices, params)

    except ImportError:
        logger.info("Quantum optimizer: dwave-neal not installed, using classical simulated annealing")

        # Classical simulated annealing fallback
        param_values = {name: _discretize_param(name) for name in PARAM_SPACE}

        # Initialize with random params
        current_params = {name: float(np.random.choice(values)) for name, values in param_values.items()}
        current_score = _objective(prices, current_params)
        best_params = current_params.copy()
        best_score = current_score

        for step in range(num_reads):
            temperature = max(0.001, 1.0 * (1 - step / num_reads))

            # Perturb one random parameter
            candidate = current_params.copy()
            param_name = np.random.choice(list(PARAM_SPACE.keys()))
            values = param_values[param_name]
            current_idx = values.index(candidate[param_name]) if candidate[param_name] in values else 0
            new_idx = max(0, min(len(values) - 1, current_idx + np.random.choice([-1, 0, 1])))
            candidate[param_name] = values[new_idx]

            candidate_score = _objective(prices, candidate)

            # Metropolis criterion
            delta = candidate_score - current_score
            if delta > 0 or np.random.random() < np.exp(delta / temperature):
                current_params = candidate
                current_score = candidate_score

            if current_score > best_score:
                best_score = current_score
                best_params = current_params.copy()

        best_result = _simulate_strategy(prices, best_params)

    return {
        "optimal_params": best_params,
        "backtest": best_result,
        "score": round(best_score, 3),
        "method": "quantum_annealing" if "neal" in str(type(best_params)) else "simulated_annealing",
    }


def quick_optimize(prices: list[float]) -> dict:
    """Quick optimization with fewer iterations for real-time use."""
    return optimize_with_annealing(prices, num_reads=100)


def full_optimize(prices: list[float]) -> dict:
    """Full optimization with more iterations for thorough search."""
    return optimize_with_annealing(prices, num_reads=500)
