requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
PyJWT[crypto]>=2.12.0
pwdlib[argon2]>=0.3.0
python-multipart>=0.0.9
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
anthropic>=0.49.0
PyNaCl>=1.5.0
pandas>=2.2.0
numpy>=1.26.0
httpx>=0.27.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
cryptography>=42.0.0
dwave-neal>=0.6.0
dimod>=0.12.0
